Refactor the player
