+++ Assets/_BForBoss/_Character/Scenes/Sandbox.unity
@@ -2093,7 +2093,7 @@ Camera:
     y: 0
     width: 1
     height: 1
-  near clip plane: 0.3
+  near clip plane: 0.2
   far clip plane: 1000
   field of view: 60
   orthographic: 0
@@ -2172,7 +2172,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   attenuationObject: {fileID: 0}
-  ListenerNumber: -1
 --- !u!4 &1280265873
 Transform:
   m_ObjectHideFlags: 0
@@ -2984,7 +2983,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 2061070007434329460, guid: c7cd698ab65444f38b1e49e2e17d8e9f, type: 3}
       propertyPath: m_AnchoredPosition.y
-      value: 0.00051098043
+      value: 0.000010657452
       objectReference: {fileID: 0}
     - target: {fileID: 3924296612012996776, guid: c7cd698ab65444f38b1e49e2e17d8e9f, type: 3}
       propertyPath: m_Name
@@ -3164,7 +3163,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 6758736603576299658, guid: c7cd698ab65444f38b1e49e2e17d8e9f, type: 3}
       propertyPath: m_AnchoredPosition.y
-      value: 0.00051098043
+      value: 0.000010657452
       objectReference: {fileID: 0}
     - target: {fileID: 7649166824825422377, guid: c7cd698ab65444f38b1e49e2e17d8e9f, type: 3}
       propertyPath: m_AnchorMax.x
@@ -3308,11 +3307,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 660e33ba81284476eb3b5b4297dce84d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _player: {fileID: 2217804276623157996}
+  _playerBehaviour: {fileID: 2217804276623157996}
   _pauseMenu: {fileID: 574067540}
   _spawnLocation: {fileID: 1621950423627791422}
   _lifeCycleManager: {fileID: 0}
-  _weaponsManager: {fileID: 0}
+  _enemyNavigationManager: {fileID: 0}
+  weaponAnimationController: {fileID: 0}
   _equipmentBehaviour: {fileID: 0}
   _ammunitionCountView: {fileID: 0}
   _reloadView: {fileID: 0}
@@ -3790,13 +3790,13 @@ PrefabInstance:
   m_SourcePrefab: {fileID: 100100000, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
 --- !u!114 &2217804276623157996 stripped
 MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7177967516030943659, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
+  m_CorrespondingSourceObject: {fileID: 3783079141729131295, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
   m_PrefabInstance: {fileID: 2217804276623157995}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 61feaf7a8fc88824d9b2669245e397a0, type: 3}
+  m_Script: {fileID: 11500000, guid: 32effb133c1fc884a9051e5c12cba2ab, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &3647294158480895873
