--- Assets/_BForBoss/_Character/Scripts/FirstPersonPlayer.Animator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca6a7bc53ac67ad4695081c341c9ee55
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
