--- Assets/_BForBoss/_Character/Scripts/FirstPersonPlayer.CameraMask.cs
@@ -1,17 +1,0 @@
-using UnityEngine;
-
-namespace Perigon.Character
-{
-    public partial class FirstPersonPlayer
-    {
-        private const string PLAYER_MODEL_LAYER = "PlayerModel";
-        private const string FIRST_PERSON_WEAPON_LAYER = "FirstPersonWeapon";
-        
-        private int FirstPersonMask => ~(1 << LayerMask.NameToLayer(PLAYER_MODEL_LAYER) | 1 << LayerMask.NameToLayer(FIRST_PERSON_WEAPON_LAYER));
-
-        private void SetCameraCullingMask()
-        {
-            camera.cullingMask = FirstPersonMask;
-        }
-    }
-}
