--- Assets/_BForBoss/_Character/Scripts/FirstPersonPlayer.CameraMask.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a3532a2e723d40c4a817b22a900d2a44
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
