--- Assets/_BForBoss/_Character/Scripts/FirstPersonPlayer.SetSpawn.cs
@@ -1,20 +1,0 @@
-using UnityEngine;
-
-namespace Perigon.Character
-{
-    public interface ICharacterSpawn
-    {
-        void SpawnAt(Vector3 position, Quaternion facing);
-    }
-
-    public partial class FirstPersonPlayer: ICharacterSpawn
-    {
-        public void SpawnAt(Vector3 position, Quaternion facing)
-        {
-            SetVelocity(Vector3.zero);
-            SetPosition(position);
-            rootPivot.rotation = facing;
-            eyePivot.rotation = facing;
-        }
-    }
-}
