--- Assets/_BForBoss/_Character/Scripts/FirstPersonPlayer.SetSpawn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ddcc5753648e74524b9b95550c36239f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
