--- Assets/_BForBoss/_Character/Scripts/FirstPersonPlayer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 61feaf7a8fc88824d9b2669245e397a0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
