+++ Assets/_BForBoss/_Character/Scripts/PlayerMovementBehaviour.Animator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca6a7bc53ac67ad4695081c341c9ee55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
