+++ Assets/_BForBoss/_Character/Scripts/PlayerMovementBehaviour.CharacterMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2136a34da59a02748b735981f58ea97a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
