+++ Assets/_BForBoss/_Character/Scripts/PlayerMovementBehaviour.InputSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb16aa037da384f0b98481e38dd84646
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
