+++ Assets/_BForBoss/_Core/SceneTemplates/WeaponRaceCourseTemplate/WeaponRaceCourse.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.33371285, g: 0.49977273, b: 0.8330078, a: 1}
+  m_IndirectSpecularColor: {r: 0.023018144, g: 0.016271964, b: 0.016159356, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -1339,7 +1339,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _firePoint: {fileID: 1467719136}
   _crosshair: {fileID: 1924278439}
-  _muzzleFlash: {fileID: 1467719138}
+  _muzzleFlash: {fileID: 0}
   _weaponScriptableObject: {fileID: 11400000, guid: b669f6e948f309b43adf276791a84958, type: 2}
 --- !u!114 &513119966
 MonoBehaviour:
@@ -1355,7 +1355,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _firePoint: {fileID: 1467719136}
   _crosshair: {fileID: 1924278439}
-  _muzzleFlash: {fileID: 1467719139}
+  _muzzleFlash: {fileID: 0}
   _weaponScriptableObject: {fileID: 11400000, guid: ebd17210773efd94da60b70929d3c5ee, type: 2}
 --- !u!114 &513119967
 MonoBehaviour:
@@ -1371,7 +1371,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _firePoint: {fileID: 1467719136}
   _crosshair: {fileID: 1924278439}
-  _muzzleFlash: {fileID: 1467719137}
+  _muzzleFlash: {fileID: 0}
   _weaponScriptableObject: {fileID: 11400000, guid: 48df86ba764786a488249b8673a3cee8, type: 2}
 --- !u!114 &513119968
 MonoBehaviour:
@@ -1847,17 +1847,6 @@ Transform:
   m_Father: {fileID: 932950509}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &907697725 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7177967516030943659, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
-  m_PrefabInstance: {fileID: 1144084895}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 61feaf7a8fc88824d9b2669245e397a0, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &932950508
 GameObject:
   m_ObjectHideFlags: 0
@@ -2212,6 +2201,17 @@ PrefabInstance:
       objectReference: {fileID: 1186343774}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
+--- !u!114 &1144084896 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3783079141729131295, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
+  m_PrefabInstance: {fileID: 1144084895}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 32effb133c1fc884a9051e5c12cba2ab, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!224 &1183948095 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 3576923979683137177, guid: 18b06456536e245408da01be69bff3e0, type: 3}
@@ -2250,7 +2250,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   attenuationObject: {fileID: 0}
-  ListenerNumber: -1
 --- !u!20 &1186343774
 Camera:
   m_ObjectHideFlags: 0
@@ -2274,7 +2273,7 @@ Camera:
     y: 0
     width: 1
     height: 1
-  near clip plane: 0.3
+  near clip plane: 0.2
   far clip plane: 1000
   field of view: 60
   orthographic: 0
@@ -2933,21 +2932,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 4917516333455242410, guid: a6233663c01e340d5a2bcaa76b1990ea, type: 3}
   m_PrefabInstance: {fileID: 1467719134}
   m_PrefabAsset: {fileID: 0}
---- !u!2083052967 &1467719137 stripped
-VisualEffect:
-  m_CorrespondingSourceObject: {fileID: 5341220295781122203, guid: a6233663c01e340d5a2bcaa76b1990ea, type: 3}
-  m_PrefabInstance: {fileID: 1467719134}
-  m_PrefabAsset: {fileID: 0}
---- !u!2083052967 &1467719138 stripped
-VisualEffect:
-  m_CorrespondingSourceObject: {fileID: 4568508669339319605, guid: a6233663c01e340d5a2bcaa76b1990ea, type: 3}
-  m_PrefabInstance: {fileID: 1467719134}
-  m_PrefabAsset: {fileID: 0}
---- !u!2083052967 &1467719139 stripped
-VisualEffect:
-  m_CorrespondingSourceObject: {fileID: 5374796527768394560, guid: a6233663c01e340d5a2bcaa76b1990ea, type: 3}
-  m_PrefabInstance: {fileID: 1467719134}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1511853182
 GameObject:
   m_ObjectHideFlags: 0
@@ -3399,7 +3383,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 3093835205963856409, guid: 1263fd774c91c9644930c6ddafd69c09, type: 3}
       propertyPath: m_AnchoredPosition.y
-      value: 0.0009813344
+      value: 0.000005609369
       objectReference: {fileID: 0}
     - target: {fileID: 4516330132463527741, guid: 1263fd774c91c9644930c6ddafd69c09, type: 3}
       propertyPath: m_AnchorMax.x
@@ -3642,12 +3626,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 58d3092f5feb2c74ab1f0b2fde65cc6f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _player: {fileID: 907697725}
+  _playerBehaviour: {fileID: 1144084896}
   _pauseMenu: {fileID: 1458297858}
   _timeManager: {fileID: 1511853183}
   _checkpointManager: {fileID: 1415184885}
   _lifeCycleManager: {fileID: 0}
-  _weaponsManager: {fileID: 932950510}
+  _weaponAnimationController: {fileID: 932950510}
   _equipmentBehaviour: {fileID: 2053475647}
   _ammunitionCountView: {fileID: 1516991103}
   _reloadView: {fileID: 2133816042}
@@ -4288,7 +4272,7 @@ MonoBehaviour:
   m_faceColor:
     serializedVersion: 2
     rgba: 4294967295
-  m_fontSize: 20.35
+  m_fontSize: 20.9
   m_fontSizeBase: 36
   m_fontWeight: 400
   m_enableAutoSizing: 1
@@ -4476,6 +4460,8 @@ GameObject:
   - component: {fileID: 2053475648}
   - component: {fileID: 2053475647}
   - component: {fileID: 2053475649}
+  - component: {fileID: 2053475651}
+  - component: {fileID: 2053475650}
   m_Layer: 0
   m_Name: Equipment
   m_TagString: Untagged
@@ -4496,6 +4482,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _inputActions: {fileID: -944628639613478452, guid: 11b5fd57f5ed74ef0a6fcf19057db823, type: 3}
+  _weaponBehaviours: []
 --- !u!4 &2053475648
 Transform:
   m_ObjectHideFlags: 0
@@ -4524,6 +4511,32 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5459813680be7f44583d74d6e70f84d1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &2053475650
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2053475646}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2479f35d5b98d6e45be90bd0e9956671, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _wallHitVFXPrefab: {fileID: 0}
+--- !u!114 &2053475651
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2053475646}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7e4392a5aea01e34ba83cc3645de52b4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _bulletPrefabs: []
 --- !u!1 &2071085758
 GameObject:
   m_ObjectHideFlags: 0
