+++ Assets/_BForBoss/_Core/SceneTemplates/WeaponSandboxSceneTemplate/WeaponSandboxSceneTemplate.unity
@@ -291,10 +291,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 660e33ba81284476eb3b5b4297dce84d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _player: {fileID: 1418917650}
+  _playerBehaviour: {fileID: 1418917650}
   _pauseMenu: {fileID: 1730176791}
   _spawnLocation: {fileID: 1154405999}
   _lifeCycleManager: {fileID: 2063217957}
+  _enemyNavigationManager: {fileID: 0}
   weaponAnimationController: {fileID: 1266760827}
   _equipmentBehaviour: {fileID: 2047271910}
   _ammunitionCountView: {fileID: 1597833375}
@@ -1505,13 +1506,13 @@ PrefabInstance:
   m_SourcePrefab: {fileID: 100100000, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
 --- !u!114 &1418917650 stripped
 MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7177967516030943659, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
+  m_CorrespondingSourceObject: {fileID: 3783079141729131295, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
   m_PrefabInstance: {fileID: 1418917649}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 61feaf7a8fc88824d9b2669245e397a0, type: 3}
+  m_Script: {fileID: 11500000, guid: 32effb133c1fc884a9051e5c12cba2ab, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1001 &1482774068
@@ -1830,7 +1831,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 2061070007434329460, guid: c7cd698ab65444f38b1e49e2e17d8e9f, type: 3}
       propertyPath: m_AnchoredPosition.y
-      value: 277.00888
+      value: 0.00097297307
       objectReference: {fileID: 0}
     - target: {fileID: 3924296612012996776, guid: c7cd698ab65444f38b1e49e2e17d8e9f, type: 3}
       propertyPath: m_Name
@@ -2010,7 +2011,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 6758736603576299658, guid: c7cd698ab65444f38b1e49e2e17d8e9f, type: 3}
       propertyPath: m_AnchoredPosition.y
-      value: 277.00888
+      value: 0.00097297307
       objectReference: {fileID: 0}
     - target: {fileID: 7649166824825422377, guid: c7cd698ab65444f38b1e49e2e17d8e9f, type: 3}
       propertyPath: m_AnchorMax.x
