+++ Assets/_BForBoss/_Core/Scripts/Debug/DebugWindow.cs
@@ -58,7 +58,7 @@ namespace BForBoss
                 }
             }
             
-            IInputSettings input = FindObjectOfType<FirstPersonPlayer>();
+            IInputSettings input = FindObjectOfType<PlayerMovementBehaviour>();
             _freezeActionsUtility = new FreezeActionsUtility(input);
         }
