+++ Assets/_BForBoss/_Core/Scripts/Managers/WeaponRaceCourseWorldManager.cs
@@ -70,13 +70,13 @@ namespace BForBoss
             _timeManager.Initialize(_timeManagerViewModel);
 
             _weaponAnimationController.Initialize(
-                () => _player.CharacterVelocity,
-                () => _player.CharacterMaxSpeed,
-                () => _player.IsWallRunning,
-                () => _player.IsGrounded,
-                () => _player.IsSliding,
-                () => _player.IsDashing);
-            _equipmentBehaviour.Initialize(_player.RootPivot);
+                () => _playerBehaviour.PlayerMovement.CharacterVelocity,
+                () => _playerBehaviour.PlayerMovement.CharacterMaxSpeed,
+                () => _playerBehaviour.PlayerMovement.IsWallRunning,
+                () => _playerBehaviour.PlayerMovement.IsGrounded,
+                () => _playerBehaviour.PlayerMovement.IsSliding,
+                () => _playerBehaviour.PlayerMovement.IsDashing);
+            _equipmentBehaviour.Initialize(_playerBehaviour.PlayerMovement.RootPivot);
             _ammunitionCountView.Initialize(_equipmentBehaviour);
             _reloadView.Initialize(_equipmentBehaviour);
