+++ Assets/_BForBoss/_Core/Scripts/PlayerBehaviour.cs
@@ -1,0 +1,37 @@
+using System;
+using Perigon.Character;
+using Perigon.Utility;
+using UnityEngine;
+
+namespace BForBoss
+{
+    [RequireComponent(typeof(PlayerMovementBehaviour))]
+    public class PlayerBehaviour : MonoBehaviour
+    {
+        private PlayerMovementBehaviour _playerMovement = null;
+
+        public PlayerMovementBehaviour PlayerMovement => _playerMovement;
+        
+        public void Initialize()
+        {
+            _playerMovement.Initialize();
+        }
+        
+        public void SpawnAt(Vector3 position, Quaternion facing)
+        {
+            _playerMovement.SetVelocity(Vector3.zero);
+            _playerMovement.SetPosition(position);
+            _playerMovement.rootPivot.rotation = facing;
+            _playerMovement.eyePivot.rotation = facing;
+        }
+
+        private void Awake()
+        {
+            _playerMovement = GetComponent<PlayerMovementBehaviour>();
+            if (_playerMovement == null)
+            {
+                PanicHelper.Panic(new Exception("PlayerMovementBehaviour missing from PlayerBehaviour"));
+            }
+        }
+    }
+}
