+++ Assets/_BForBoss/_LevelDesign/Scenes/MS2_RaceCourse.unity
@@ -10037,17 +10037,6 @@ Mesh:
     offset: 0
     size: 0
     path: 
---- !u!114 &907697725 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7177967516030943659, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
-  m_PrefabInstance: {fileID: 1144084895}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 61feaf7a8fc88824d9b2669245e397a0, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!43 &924186777
 Mesh:
   m_ObjectHideFlags: 0
@@ -12427,6 +12416,17 @@ PrefabInstance:
       objectReference: {fileID: 1186343774}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
+--- !u!114 &1144084896 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3783079141729131295, guid: 9cef2f6c7c3da6449a278b7a673ecbd2, type: 3}
+  m_PrefabInstance: {fileID: 1144084895}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 32effb133c1fc884a9051e5c12cba2ab, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1178930636
 GameObject:
   m_ObjectHideFlags: 0
@@ -17241,7 +17241,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 3093835205963856409, guid: 1263fd774c91c9644930c6ddafd69c09, type: 3}
       propertyPath: m_AnchoredPosition.y
-      value: 0.00034068237
+      value: 0.0007637413
       objectReference: {fileID: 0}
     - target: {fileID: 4516330132463527741, guid: 1263fd774c91c9644930c6ddafd69c09, type: 3}
       propertyPath: m_AnchorMax.x
@@ -17712,7 +17712,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 58d3092f5feb2c74ab1f0b2fde65cc6f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _player: {fileID: 907697725}
+  _playerBehaviour: {fileID: 1144084896}
   _pauseMenu: {fileID: 1458297858}
   _timeManager: {fileID: 1511853183}
   _checkpointManager: {fileID: 1415184885}
